+++ ui/components/ItemDetail/ItemDetail.js
@@ -1,5 +1,4 @@
-import React, { useState, useEffect } from "react";
-import { useSelector } from "react-redux";
+import React, { useState, useEffect, useContext } from "react";
 import PeJobDetail from "./PeJobDetail";
 import MatchaDetail from "./MatchaDetail";
 import LbbCompanyDetail from "./LbbCompanyDetail";
@@ -13,7 +12,7 @@ import { capitalizeFirstLetter } from "../../utils/strutils";
 import { isCfaEntreprise } from "../../services/cfaEntreprise";
 import { filterLayers } from "../../utils/mapTools";
 import ExternalLink from "../externalLink";
-
+import { SearchResultContext } from "../../context/SearchResultContextProvider";
 import { useSwipeable } from "react-swipeable";
 import { mergeJobs, mergeOpportunities } from "../../utils/itemListUtils";
 
@@ -61,6 +60,23 @@ const getActualTitle = ({ selectedItem, kind }) => {
   return title;
 };
 
+const getCurrentList = (store, activeFilter, extendedSearch) => {
+  let picked = pick(store, ["trainings", "jobs"]);
+  let trainingsArray = amongst(activeFilter, ["all", "trainings"]) ? get(picked, "trainings", []) : [];
+
+  let jobList = [];
+  let companyList = [];
+  if (amongst(activeFilter, ["all", "jobs"])) {
+    if (extendedSearch) jobList = mergeOpportunities(get(picked, "jobs"));
+    else {
+      jobList = mergeJobs(get(picked, "jobs"));
+      companyList = mergeOpportunities(get(picked, "jobs"), "onlyLbbLba");
+    }
+  }
+  let fullList = concat([], trainingsArray, jobList, companyList);
+return  fullList.filter((el) => !!el);
+};
+
 const ItemDetail = ({ selectedItem, handleClose, displayNavbar, handleSelectItem, activeFilter }) => {
   const kind = selectedItem?.ideaType;
 
@@ -80,25 +96,8 @@ const ItemDetail = ({ selectedItem, handleClose, displayNavbar, handleSelectItem
 
   let actualTitle = getActualTitle({ kind, selectedItem });
 
-  const { extendedSearch, formValues } = useSelector((state) => state.trainings);
-
-  const currentList = useSelector((store) => {
-    let picked = pick(store.trainings, ["trainings", "jobs"]);
-    let trainingsArray = amongst(activeFilter, ["all", "trainings"]) ? get(picked, "trainings", []) : [];
-
-    let jobList = [];
-    let companyList = [];
-    if (amongst(activeFilter, ["all", "jobs"])) {
-      if (extendedSearch) jobList = mergeOpportunities(get(picked, "jobs"));
-      else {
-        jobList = mergeJobs(get(picked, "jobs"));
-        companyList = mergeOpportunities(get(picked, "jobs"), "onlyLbbLba");
-      }
-    }
-    let fullList = concat([], trainingsArray, jobList, companyList);
-    let listWithoutEmptyValues = fullList.filter((el) => !!el);
-    return listWithoutEmptyValues;
-  });
+  const { trainings, jobs, extendedSearch } = useContext(SearchResultContext);
+  const currentList = getCurrentList({ trainings, jobs }, activeFilter, extendedSearch);
 
   // See https://www.npmjs.com/package/react-swipeable
   const swipeHandlers = useSwipeable({