+++ ui/components/ItemDetail/LbbCompany.js
@@ -1,7 +1,6 @@
 import React, { useState } from "react";
 import jobIcon from "../../public/images/icons/job.svg";
 import TagCandidatureSpontanee from "./TagCandidatureSpontanee";
-import { useSelector } from "react-redux";
 import { fetchAddresses } from "../../services/baseAdresse";
 import extendedSearchPin from "../../public/images/icons/trainingPin.svg";
 import { capitalizeFirstLetter } from "../../utils/strutils";
@@ -9,9 +8,12 @@ import { get } from "lodash";
 import { setSelectedMarker } from "../../utils/mapTools";
 import { getItemQueryParameters } from "../../utils/getItemId";
 import { getSearchQueryParameters } from "../../utils/getSearchParameters";
+import { SearchResultContext } from "../../context/SearchResultContextProvider";
+import { DisplayContext } from "../../context/DisplayContextProvider";
 
 const LbbCompany = ({ company, handleSelectItem, showTextOnly, searchForTrainingsOnNewCenter }) => {
-  const { formValues, selectedMapPopupItem } = useSelector((state) => state.trainings);
+  const { selectedMapPopupItem } = React.useContext(SearchResultContext);
+  const { formValues } = React.useContext(DisplayContext);
 
   const currentSearchRadius = formValues?.radius || 30;