+++ ui/components/Map/MapSearchButton.js
@@ -1,10 +1,13 @@
 import React from "react";
-import { useSelector } from "react-redux";
+
 import refreshSearchOnMap from "public/images/icons/refreshSearchOnMap.svg";
 import lookingGlassOnMap from "public/images/glass.svg";
+import { SearchResultContext } from "../../context/SearchResultContextProvider";
+import { DisplayContext } from "../../context/DisplayContextProvider";
 
 const MapSearchButton = ({ handleSearchClick }) => {
-  const { hasSearch, formValues } = useSelector((state) => state.trainings);
+  const { formValues } = React.useContext(DisplayContext);
+  const { hasSearch } = React.useContext(SearchResultContext);
 
   return hasSearch ? (
     <div className="c-map-searchButton">