+++ ui/components/SearchForTrainingsAndJobs/components/NoJobResult.js
@@ -1,8 +1,8 @@
-import React from "react";
-import { useScopeContext } from "../../../context/ScopeContext";
+import React, { useContext } from "react";
+import { ScopeContext } from "../../../context/ScopeContext";
 
 const NoJobResult = () => {
-  const scopeContext = useScopeContext();
+  const scopeContext = useContext(ScopeContext);
 
   return scopeContext.isJob ? <div className="bold">Aucune entreprise trouvée pour votre recherche</div> : "";
 };