+++ ui/components/footer.js
@@ -1,15 +1,10 @@
 import React from "react";
 import { Row, Col, Container } from "reactstrap";
 import ExternalLink from "./externalLink";
-import { push } from "connected-next-router";
-import { useDispatch } from "react-redux";
+import { useRouter } from "next/router";
 
-/*
- Different kind of navigation are available here :
- https://raw.githubusercontent.com/danielr18/connected-next-router/master/examples/basic/components/navigation.js
-*/
 const Footer = (props) => {
-  const dispatch = useDispatch();
+  const router = useRouter();
 
   return (
     <>
@@ -60,7 +55,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword pr-3"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/mentions-legales" }));
+                      router.push("/mentions-legales");
                     }}
                     href="/mentions-legales"
                   >
@@ -73,7 +68,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/cgu" }));
+                      router.push("/cgu");
                     }}
                     href="/cgu"
                   >
@@ -86,7 +81,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/cookies" }));
+                      router.push("/cookies");
                     }}
                     href="/cookies"
                   >
@@ -99,7 +94,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/stats" }));
+                      router.push("/stats");
                     }}
                     href="/stats"
                   >
@@ -112,7 +107,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/faq" }));
+                      router.push("/faq");
                     }}
                     href="/faq"
                   >
@@ -125,7 +120,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/contact" }));
+                      router.push("/contact");
                     }}
                     href="/contact"
                   >
@@ -138,7 +133,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/metiers" }));
+                      router.push("/metiers");
                     }}
                     href="/metiers"
                   >
@@ -151,7 +146,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/a-propos" }));
+                      router.push("/a-propos");
                     }}
                     href="/a-propos"
                   >
@@ -164,7 +159,7 @@ const Footer = (props) => {
                     className="c-footer-links__link c-footer-smallword"
                     onClick={(e) => {
                       e.preventDefault();
-                      dispatch(push({ pathname: "/developpeurs" }));
+                      router.push("/developpeurs");
                     }}
                     href="/developpeurs"
                   >