+++ server/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 1.19.0 [22/05/2022]
+[LBA6-68] Remplacement redux par API context
+[LBA6-36] Fix effect de rechargement sur lien Organisme de formation
+
 ## 1.18.12 [17/05/2022]
 [LBA6-132] Corrections en-têtes des fiches détails
 [LBA6-131] Scrolling vers les offres cassé